requirements.txt
async-to-sync==0.2.2
bleak==0.13.0
